+++ Sources/Collection+Utility.swift
@@ -6,11 +6,14 @@ extension MutableCollection {
     /// Reorders the elements of a collection such that all the elements
     /// that match a given predicate are after all the elements that don't match.
     ///
-    /// - Parameter belongsInSecondPartition: A predicate used to partition the collection.
-    ///   All elements satisfying this predicate are ordered after all elements not satisfying it.
+    /// - Parameters:
+    ///    - belongsInSecondPartition: A predicate used to partition the collection.
+    ///      All elements satisfying this predicate are ordered after all elements not satisfying it.
+    ///    - element: An element to test
+    ///
     /// - Returns: Slices representing the matching and non-matching partitions.
-    public mutating func partitionSplit(by belongsInSecondPartition: (Self.Element) throws -> Bool) rethrows -> (matching: SubSequence, notMatching: SubSequence) {
+    public mutating func partitionSplit(by belongsInSecondPartition: (_ element: Self.Element) throws -> Bool) rethrows -> (matching: SubSequence, notMatching: SubSequence) {
         let idx = try self.partition(by: belongsInSecondPartition)
-        return (matching: self[..<idx], notMatching: self[idx...])
+        return (matching: self[idx...], notMatching: self[..<idx])
     }
 }