+++ Sources/Result+Utility.swift
@@ -1,6 +1,6 @@
 //  Copyright © 2020 Erica Sadun. All rights reserved.
 
-public extension Result {
+extension Result {
     /// Initializes a `Result` from a completion handler's `(data?, error?)`.
     ///
     /// When both data and error are non-nil, `Result` first populates the
@@ -9,7 +9,7 @@ public extension Result {
     /// - Parameters:
     ///   - data: the optional data returned via a completion handler
     ///   - error: the optional error returned via a completion handler
-    init(_ data: Success?, _ error: Failure?) {
+    public init(_ data: Success?, _ error: Failure?) {
         precondition(!(data == nil && error == nil))
         switch (data, error) {
         case (_, let failure?): self = .failure(failure)