+++ Sources/SynchronousData.swift
@@ -2,13 +2,13 @@
 
 import Foundation
 
-public extension Utility {
+extension Utility {
     
-    /// Creates a blocking URLSession data task
-    /// - Parameter request: the `URLRequest` for the session
-    /// - Throws: Any returned error on `Result.get`
-    /// - Returns: The retrieved session data
-    static func requestSynchronousData(_ request: URLRequest) throws -> Data {
+    /// Creates a blocking URLSession data task.
+    /// - Parameter request: the `URLRequest` for the session.
+    /// - Throws: Any returned error on `Result.get`.
+    /// - Returns: The retrieved session data.
+    public static func requestSynchronousData(_ request: URLRequest) throws -> Data {
         let semaphore = DispatchSemaphore(value: 0)
         var result: Result<Data, Error> = Result(Data(), nil)
         let task = URLSession.shared.dataTask(with: request) { data, _, error -> Void in