+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.0.3
+
+Add: FileManager touch (url and string path)
+Add: URL versioning
+
+Added tests
+
+Refactored Mac-specific code out to Mac-General-Utility
+
 ## 0.0.2
 
 Add: isDir checks (FileManger, String), mutable partition slices